+++ photopedia/src/Components/AddLocations.css
@@ -1,0 +1,96 @@
+body {
+    width: 100%;
+    height: auto;
+}
+
+.header_text_location {
+    display: flex;
+    justify-content: center;
+    text-align: center;
+}
+
+.header_text_location p {
+    display: flex;
+    justify-content: center;
+    font-size: 20px;
+}
+
+
+.form_container {
+    width: 33%;
+    height: 65vh;
+    background-image: url('./../images/add_locations_640x640.jpg');
+    background-repeat: no-repeat;
+    border-radius: 10px;
+    display: flex;
+    justify-content: center;
+    margin-left: 640px;
+    margin-top: 50px;
+} 
+
+.form_background {
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.267);
+    opacity: 0.7;
+    border-radius: 10px;
+    display: flex;
+    justify-content: center;
+    /* align-items: center; */
+    text-align: center;
+    padding-top: 20px;
+}
+
+.form_inputs {
+    width: 70%;
+    display: flex;
+    justify-content: center;
+    text-align: center;
+    border-radius: 10px;
+    
+}
+
+form {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
+    font-size: 18px;
+    width: 100%;
+    /* border: none;
+    align-items: center;
+    text-align: center; */
+    
+}
+
+input {
+    background-color: rgb(0, 0, 0);
+    border:none;
+    border-radius: 5px;
+    color: white;
+}
+
+button {
+    width: 120px;
+    height: 60px;
+    border-radius: 5px;
+    border: none;
+    text-align: center;
+    /* align-self: center; */
+    background-color: rgb(0, 0, 0);
+    color: white;
+}
+
+button:hover {
+    background-color: rgba(0, 0, 0, 0.329);
+    color: black;
+}
+
+.social_media_addlocations {
+    display: flex;
+    justify-content: center;
+}
+
+.location_button {
+    display: flex;
+    justify-content: space-between;
+}
